moved chembl_db to db_generation folder
