+++ db_generation/chembl_db/download_parse_ccd.py
