+++ db_generation/chembl_db/uniprot_pfam.py
